requirements.txt
# Core
opencv-python==4.7.0.72
numpy==1.23.5
pandas==1.5.3
matplotlib

# PyTorch (CUDA-enabled if available)
torch
torchvision
torchaudio

# Object detection (YOLOv8)
ultralytics==8.0.196

# Face recognition
deepface==0.0.79
onnxruntime==1.15.1

# Voice input/output
sounddevice==0.4.6
soundfile==0.12.1
pyttsx3==2.90
pywin32==306
playsound

# Speech-to-text
openai-whisper==20231117

# LLM (local AI brain)
llama-cpp-python==0.2.78

# Utils
requests==2.31.0
